--- a/rspack.config.ts
+++ b/rspack.config.ts
@@ -284,6 +284,10 @@ const appConfig: Configuration = {
     // Assets path should be `../assets/rubik.woff` not `assets/rubik.woff`
     // Not compatible with CssExtractRspackPlugin https://rspack.rs/guide/tech/css#using-cssextractrspackplugin
     css: false,
+    // https://rspack.dev/config/experiments#experimentslazybarrel
+    lazyBarrel: true,
+    // https://rspack.dev/config/experiments#experimentsnativewatcher
+    nativeWatcher: true,
   },
   module: {
     /**
@@ -550,19 +554,7 @@ const appConfig: Configuration = {
     // This only runs in production mode
     minimizer: [
       new rspack.LightningCssMinimizerRspackPlugin(),
-      new rspack.SwcJsMinimizerRspackPlugin({
-        minimizerOptions: {
-          compress: {
-            // We are turning off these 3 minifier options because it has caused
-            // unexpected behaviour. See the following issues for more details.
-            // - https://github.com/swc-project/swc/issues/10822
-            // - https://github.com/swc-project/swc/issues/10824
-            reduce_vars: false,
-            inline: 0,
-            collapse_vars: false,
-          },
-        },
-      }),
+      new rspack.SwcJsMinimizerRspackPlugin(),
     ],
   },
   devtool: IS_PRODUCTION ? 'source-map' : 'eval-cheap-module-source-map',