https://rspack.rs/blog/announcing-1-5

enables two experiemental features that mostly affect local dev. Re-enables swc inlining since the issues were closed.

Decent speed up in HMR. Unscientific measurements of a few file edits

before
```
HMR1 546ms
HMR2 529ms
```

after
```
HMR1 424ms
HMR2 403ms
```